* Make unixfilesystemhandler less verbose, and the verbosity that it still has   is now written to CLogOutput (infolog.txt) instead of stderr. * Added __attribute__'s to CLogOutput to make gcc give warnings about invalid format strings. * Fixed some gcc warnings. * Added 'Any++ speedup' default keybinding.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@2217 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Rendering/Textures/ColorMap.cpp
@@ -76,7 +76,7 @@ unsigned char* CColorMap::GetColor(unsigned char *color, float pos)
 	float fposn = pos*nxsize;
 	int iposn = (int)fposn;
 	float fracn = fposn-iposn;
-	int aa = fracn*256;
+	int aa = int(fracn*256);
 	int ia = 256-aa;
 
 	unsigned char *col1 = (unsigned char*)&map[iposn*4];
@@ -103,7 +103,7 @@ void CColorMap::LoadFromFloatVector(std::vector<float> &vec)
 
 	for(int i=0; i<xsize*4; i++)
 	{
-		lmap[i] = vec[i]*255;
+		lmap[i] = int(vec[i]*255);
 	}
 	LoadMap(lmap, xsize*4);
 	delete [] lmap;
